define SymPy extension
